+++ Editor/Drawers/SaintsDictionary/SaintsDictionaryDrawerUIToolkit.cs
@@ -1,4 +1,4 @@
-#if UNITY_2022_2_OR_NEWER && !SAINTSFIELD_DEBUG_UNITY_FUCKED_UP
+#if UNITY_2022_2_OR_NEWER && !SAINTSFIELD_DEBUG_UNITY_BROKEN_FALLBACK
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -798,7 +798,7 @@ namespace SaintsField.Editor.Drawers.SaintsDictionary
         }
     }
 }
-#elif UNITY_2021_3_OR_NEWER || SAINTSFIELD_DEBUG_UNITY_FUCKED_UP
+#elif UNITY_2021_3_OR_NEWER || SAINTSFIELD_DEBUG_UNITY_BROKEN_FALLBACK
 using System.Collections.Generic;
 using System.Reflection;
 using SaintsField.Editor.Utils;